+++ oauth-server/src/utils/base64.js
@@ -1,0 +1,3 @@
+export function base64url(base64Str) {
+  return base64Str.replace(/=/g, '').replace(/\+/g, '-').replace(/\//g, '_');
+}