+++ src/steps/account/index.test.ts
@@ -1,9 +1,8 @@
 import { executeStepWithDependencies } from '@jupiterone/integration-sdk-testing';
 import { buildStepTestConfigForStep } from '../../../test/config';
 import { Recording, setupProjectRecording } from '../../../test/recording';
-import { Steps } from '../constants';
+import { IntegrationSteps } from '../constants';
 
-// See test/README.md for details
 let recording: Recording;
 afterEach(async () => {
   await recording.stop();
@@ -15,7 +14,7 @@ test('fetch-account', async () => {
     name: 'fetch-account',
   });
 
-  const stepConfig = buildStepTestConfigForStep(Steps.ACCOUNT);
+  const stepConfig = buildStepTestConfigForStep(IntegrationSteps.ACCOUNT);
   const stepResult = await executeStepWithDependencies(stepConfig);
   expect(stepResult).toMatchStepMetadata(stepConfig);
 });