+++ src/steps/project-membership/index.test.ts
@@ -1,43 +1,48 @@
 import { executeStepWithDependencies } from '@jupiterone/integration-sdk-testing';
 import { buildStepTestConfigForStep } from '../../../test/config';
 import { Recording, setupProjectRecording } from '../../../test/recording';
-import { Steps } from '../constants';
+import { IntegrationSteps } from '../constants';
 
-// See test/README.md for details
 let recording: Recording;
 afterEach(async () => {
   await recording.stop();
 });
 
-test('fetch-users', async () => {
+test('fetch-and-build-project-memberships', async () => {
   recording = setupProjectRecording({
     directory: __dirname,
-    name: 'fetch-users',
+    name: 'fetch-and-build-project-memberships',
   });
 
-  const stepConfig = buildStepTestConfigForStep(Steps.USERS);
+  const stepConfig = buildStepTestConfigForStep(
+    IntegrationSteps.FETCH_AND_BUILD_PROJECT_MEMBERSHIPS,
+  );
   const stepResult = await executeStepWithDependencies(stepConfig);
   expect(stepResult).toMatchStepMetadata(stepConfig);
 });
 
-test('fetch-groups', async () => {
+test('build-project-membership-and-user-relationship', async () => {
   recording = setupProjectRecording({
     directory: __dirname,
-    name: 'fetch-groups',
+    name: 'build-project-membership-and-user-relationship',
   });
 
-  const stepConfig = buildStepTestConfigForStep(Steps.GROUPS);
+  const stepConfig = buildStepTestConfigForStep(
+    IntegrationSteps.BUILD_PROJECT_MEMBERSHIP_AND_USER_RELATIONSHIP,
+  );
   const stepResult = await executeStepWithDependencies(stepConfig);
   expect(stepResult).toMatchStepMetadata(stepConfig);
 });
 
-test('build-user-group-relationships', async () => {
+test('build-project-membership-allows-project-relationship', async () => {
   recording = setupProjectRecording({
     directory: __dirname,
-    name: 'build-user-group-relationships',
+    name: 'build-project-membership-allows-project-relationship',
   });
 
-  const stepConfig = buildStepTestConfigForStep(Steps.GROUP_USER_RELATIONSHIPS);
+  const stepConfig = buildStepTestConfigForStep(
+    IntegrationSteps.BUILD_PROJECT_MEMBERSHIP_ALLOWS_PROJECT_RELATIONSHIP,
+  );
   const stepResult = await executeStepWithDependencies(stepConfig);
   expect(stepResult).toMatchStepMetadata(stepConfig);
 });